Initial commit
With example images; including an ancient SSD.

Setting this up is easy if you know a bit of docker/nginx. Otherwise
your local chatbot is probably willing to help.

Bugs: you may need to take pictures in the right light with the right
angle. It's far from smart. But it gets the jobs done.